+++ packages/sdk/test/ci/plt/TokenAmount.test.ts
@@ -14,6 +14,16 @@ describe('PLT TokenAmount', () => {
         );
     });
 
+    test('Parses large decimal values correctly', () => {
+        expect(() => TokenAmount.fromDecimal('1', 255)).toThrow(TokenAmount.Err.exceedsMaxValue());
+        expect(
+            TokenAmount.fromDecimal(
+                '0.000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001',
+                255
+            )
+        ).toEqual(TokenAmount.create(1n, 255));
+    });
+
     test('Token amounts with invalid values throws', () => {
         expect(() => TokenAmount.create(-504n, 0)).toThrow(TokenAmount.Err.negative());
         expect(() => TokenAmount.create(MAX_U64 + 1n, 0)).toThrow(TokenAmount.Err.exceedsMaxValue());